Move the glue code to execution_manager.cpp.
This throws in a weak definition of cudaq::getExecutionManager() to let
everything link in libcudaq.so. In order to really use this however, a
proper execution manager must be linked to override the weak definition.
